+++ tests/testthat/test-scaffold.R
@@ -6,3 +6,10 @@ test_that("renv.lock is created when scaffold is called", {
   scaffold()
   expect_true(file.exists("renv.lock"))
 })
+
+test_that("renv::scaffold() accepts project settings", {
+  renv_tests_scope()
+  scaffold(settings = list(ignored.packages = "A"))
+  ignored <- settings$ignored.packages()
+  expect_identical(ignored, "A")
+})